Move LspOpenLocationCommand to core/registry.py
